LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-insets-grid-expected.txt

abcdef

FAIL Position absolute getComputedStyle left (for display: grid) assert_equals: Invalid gCS($("body > span > span"))['left']; expected "0px" but got "100px"
